requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
blis==0.2.4
catalogue==1.0.0
certifi==2020.4.5.2
chardet==3.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==7.1.2
cymem==2.0.3
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.0
packaging==20.4
pathspec==0.8.0
Pint==0.13
plac==0.9.6
preshed==2.0.1
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8
pytz==2020.1
PyYAML==3.13
regex==2020.1.7
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
spacy==2.1.9
SQLAlchemy==1.3
srsly==1.0.2
thinc==7.0.8
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.1.0
pytest==6.2.5
wikipedia==1.4.0
pdfminer.six==20211012
pylint==2.12.2
PyMuPDF==1.19.5
PyPDF2==1.26.0